Fix current buffer copying in `enter_mode!`
The two argument version of `enter_mode!` is intended to allow REPL modes to
work like Julia's builtin REPL modes: when there is already partial input on
the prompt, and the user switches the mode, then the partial input is carried
forward to the new REPL mode.

However, the copy was done too late, in default_trans_action() instead of
before that closure is created and control handed off to LineEdit. This seems
to have been broken since a refactoring in August 2019.